refactor : 노선 등록 기능 코드 리팩토링
- views 폴더로 SectionManager 파일 이동
- 불필요한 변수 삭제 및 빠른 return
- 메소드 순서 변경
