+++ src/components/template/sectionManagerTemplate.js
@@ -1,4 +1,4 @@
-import stationStorage from '../../utils/stationStorage.js';
+import stationStorage from '../../storage/stationStorage.js';
 
 const SECTION_MANAGER_PAGE_MENU_TEMPLATE = `<section class="section-buttons-container">
     <h3>구간을 수정할 노선을 선택해주세요</h3>
@@ -6,6 +6,10 @@ const SECTION_MANAGER_PAGE_MENU_TEMPLATE = `<section class="section-buttons-cont
   </section>
   <section id="section-selector-container"></section>
   <section id="section-table"></section>`;
+
+const SECTION_LINE_MENU_BTN = (line) => {
+  return `<button class="section-line-menu-button" data-lineId=${line.id}>${line.name}</button>`;
+};
 
 const SECTION_MANAGER_PAGE_SELECTOR_TEMPLATE = `<h3 id="line-title"></h3>
   <h4>구간 등록</h4>
@@ -37,10 +41,6 @@ const SECTION_TABLE_TEMPLATE = ({ id, stationIds }) => {
        </tr>`
     )
     .join('');
-};
-
-const SECTION_LINE_MENU_BTN = (line) => {
-  return `<button class="section-line-menu-button" data-lineId=${line.id}>${line.name}</button>`;
 };
 
 export {
